requirements_dev.txt
dvc
pandas
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
mlflow==1.26.1
urllib3
-e .

## Dev requirements
pytest==7.1.3
tox==3.25.1
black==22.8.0
flake8==5.0.4
mypy==0.971
